requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-swagger==1.0.16
aiosignal==1.2.0
arrow==1.2.3
async-lru==1.0.3
async-timeout==4.0.2
attrs==22.1.0
autoflake==1.7.6
autoflake8==0.2.2
backcall==0.2.0
backports.ssl==0.0.9
bandit==1.7.1
bcrypt==4.0.1
binaryornot==0.4.4
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cisco-gnmi==1.0.16
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cookiecutter==2.1.1
cryptography==38.0.1
debugpy==1.5.1
decorator==5.1.0
dill==0.3.5.1
distro==1.8.0
entrypoints==0.3
eradicate==2.0.0
flake8==4.0.1
flake8-builtins==1.5.3
flake8-eradicate==1.2.0
flake8-print==4.0.0
frozenlist==1.3.1
genie==22.9
genie.libs.clean==22.9.1
genie.libs.conf==22.9
genie.libs.filetransferutils==22.9
genie.libs.health==22.9
genie.libs.ops==22.9
genie.libs.parser==22.9
genie.libs.robot==22.9
genie.libs.sdk==22.9.1
genie.telemetry==22.9
genie.trafficgen==22.9
gitdb==4.0.9
GitPython==3.1.18
grpcio==1.50.0
idna==3.4
ipykernel==6.6.0
ipython==7.30.1
isort==5.10.1
IxNetwork==9.0.1915.16
ixnetwork-restpy==1.1.6
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==2.2.0
junit-xml==1.9
jupyter-client==7.1.0
jupyter-core==4.9.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.2
mypy==0.931
mypy-extensions==0.4.3
ncclient==0.6.13
nest-asyncio==1.5.4
netaddr==0.8.0
packaging==21.3
paramiko==2.11.0
parso==0.8.3
pathspec==0.10.1
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
platformdirs==2.4.0
prettytable==3.4.1
prompt-toolkit==3.0.23
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pyats==22.9
pyats.aereport==22.9
pyats.aetest==22.9
pyats.async==22.9
pyats.connections==22.9
pyats.contrib==22.9
pyats.datastructures==22.9
pyats.easypy==22.9
pyats.kleenex==22.9
pyats.log==22.9
pyats.reporter==22.9
pyats.results==22.9
pyats.robot==22.9
pyats.tcl==22.9
pyats.topology==22.9
pyats.utils==22.9
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pyftpdlib==1.5.7
Pygments==2.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==3.14.2
python-slugify==6.1.2
python-socketio==4.6.1
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.8
requests==2.28.1
requests-toolbelt==0.10.0
rich==12.6.0
robotframework==5.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==3.5.0
text-unidecode==1.3
tftpy==0.8.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.4.0
unicon==22.9
unicon.plugins==22.9
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.1
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
xmltodict==0.13.0
yamllint==1.26.3
yang.connector==22.9
yarl==1.8.1
